Separate room description into a component
Issue #27
